Crate governance_os_support

A compilation of traits and helpers for implementing the Governance OS

Re-exports

pub use currencies::Currencies;

pub use currencies::ReservableCurrencies;

Modules

acl
A set of common traits to define Access Control lists between pallets and runtime users.

currencies
A collection of traits and helpers to create a runtime that can support multiple currencies by implementing a trait similar to Currency. We tried to maintain API compatibility between the native Currency and this one.

testing
A series of helpers and common values used for tests.

Macros

impl_enum_default
Use this macro to easily implement Default for a given enum. This avoids having to type the same code everytime.

mock_runtime
This macro makes it easy to prepare a mock testing runtime. It also avoid the developer from duplicating mocking code. However, the developer is still expected to write the correct imports for now. The mock runtime is composed of the system and bylaws pallets as all the other pallets of the Governance OS project are supposed to depend on those; so better not write the same code twice.