Module governance_os_support::traits

Enums

ProposalResult
End result of a proposal being closed.

Traits

Currencies
Abstraction trait over a multiple currencies system, each currency type is identified by a CurrencyId, if it is set to None when calling a function the implementation should default to the native currency of the runtime.

LockableCurrencies
An extension of the Currencies trait to allow the runtime to lock funds from the token holders. Locks should be combinable. Creating a lock shall increment any existing reference count, deleting one should decrement it.

ReservableCurrencies
An extension of the Currencies trait to allow the runtime to reserve funds from the token holders.

RoleManager
This trait can be implemented by a pallet to expose an interface for other pallets to manage their own role based access control features.

StandardizedVoting
A common trait accross all voting implementations to make it easy to change between voting models or implementations. A pallet implementing this trait is not necessarily in charge of storing proposals but could stick to only to support the actual decison making code while proposal storage is delegated to another pallet.

VotingRouter
Used to route votes and related actions between different voting system implementations.