Trait governance_os_support::traits::VotingHooks

pub trait VotingHooks {
    type AccountId;
    type BlockNumber;
    type Currencies: ReservableCurrencies<Self::AccountId>;
    type Data;
    type OrganizationId;
    type VotingSystem;
    fn on_create_proposal(
        voting_system: Self::VotingSystem, 
        creator: &Self::AccountId, 
        current_block: Self::BlockNumber
    ) -> (DispatchResult, Self::Data);

    fn on_veto_proposal(
        voting_system: Self::VotingSystem, 
        data: Self::Data
    ) -> DispatchResult;

    fn on_decide_on_proposal(
        voting_system: Self::VotingSystem, 
        data: Self::Data, 
        voter: &Self::AccountId, 
        power: <Self::Currencies as Currencies<Self::AccountId>>::Balance, 
        in_support: bool
    ) -> (DispatchResult, Self::Data);

    fn can_close(
        voting_system: Self::VotingSystem, 
        data: Self::Data, 
        current_block: Self::BlockNumber
    ) -> bool;

    fn passing(voting_system: Self::VotingSystem, data: Self::Data) -> bool;

    fn on_close_proposal(
        voting_system: Self::VotingSystem, 
        data: Self::Data, 
        executed: bool
    ) -> DispatchResult;
}

Called by the host pallet to let the developer implement custom voting actions according to its own model.

Associated Types

type AccountId

type BlockNumber

type Currencies: ReservableCurrencies<Self::AccountId>

type Data

type OrganizationId

type VotingSystem

Loading content...
Required methods

fn on_create_proposal(
voting_system: Self::VotingSystem,
creator: &Self::AccountId,
current_block: Self::BlockNumber
) -> (DispatchResult, Self::Data)

Somebody is creating a proposal. Called before any state changes. This is where you have the ability to try and reserve a certain amount of coins for instance.

fn on_veto_proposal(
voting_system: Self::VotingSystem,
data: Self::Data
) -> DispatchResult

A proposal is going to be vetoed. Called before any state changes. This is where you have the possibility to free any funds reserved.

fn on_decide_on_proposal(
voting_system: Self::VotingSystem,
data: Self::Data,
voter: &Self::AccountId,
power: <Self::Currencies as Currencies<Self::AccountId>>::Balance,
in_support: bool
) -> (DispatchResult, Self::Data)

Handle an incoming vote. Returns a result and some updated metadata.

fn can_close(
voting_system: Self::VotingSystem,
data: Self::Data,
current_block: Self::BlockNumber
) -> bool

Return wether we should enable calls to closing the proposal. Closing a proposal means executing it if it passed and then cleaning the storage.

fn passing(voting_system: Self::VotingSystem, data: Self::Data) -> bool

Return if a proposal is passing. If it is, it will likely be executed in the same transaction and then closed.

fn on_close_proposal(
voting_system: Self::VotingSystem,
data: Self::Data,
executed: bool
) -> DispatchResult

Called before cleaning the storage related to a proposal and before its eventual execution. Last chance to prevent it from running!

Loading content...